+++ source/graphics/Program.cpp
@@ -32,7 +32,7 @@ void Program::create()
 
 void Program::attach(GLenum type, const GLchar* source)
 {
-    GLint length = strlen(source);
+    GLint length = GLint(strlen(source));
 
     auto shader = glCreateShader(type);
     glShaderSource(shader, 1, &source, &length);
