+++ source/graphics/util.cpp
@@ -8,7 +8,7 @@ void loadMat3ToUniform(const Mat3& mat, GLint location)
 
     for(auto i = 0; i < 9; i++)
     {
-        m[i] = mat.m[i];
+        m[i] = GLfloat(mat.m[i]);
     }
 
     glUniformMatrix3fv(location, 1, GL_FALSE, m);
@@ -20,7 +20,7 @@ void loadMat4ToUniform(const Mat4& mat, GLint location)
 
     for(auto i = 0; i < 16; i++)
     {
-       m[i] = mat.m[i];
+       m[i] = GLfloat(mat.m[i]);
     }
 
     glUniformMatrix4fv(location, 1, GL_FALSE, m);
